+++ tests/drawing_test.py
@@ -1,20 +1,12 @@
 import itertools
 
-from pytest import fixture, mark, raises, skip
+from pytest import mark, raises
 
 from wand.image import Image
 from wand.color import Color
 from wand.api import library
 from wand.drawing import Drawing
-from wand.exceptions import WandLibraryVersionError
 from wand.version import MAGICK_VERSION_NUMBER
-
-
-@fixture
-def fx_wand(request):
-    wand = Drawing()
-    request.addfinalizer(wand.destroy)
-    return wand
 
 
 def test_init_user_error():
@@ -23,229 +15,255 @@ def test_init_user_error():
             pass
 
 
-def test_is_drawing_wand(fx_wand):
-    assert library.IsDrawingWand(fx_wand.resource)
+def test_is_drawing_wand():
+    with Drawing() as ctx:
+        assert library.IsDrawingWand(ctx.resource)
 
 
-def test_set_get_border_color(fx_wand):
-    with Color("#0F0") as green:
-        fx_wand.border_color = green
-        assert green == fx_wand.border_color
-    fx_wand.border_color = 'orange'
-    assert fx_wand.border_color == Color('orange')
-    # Assert user error
-    with raises(TypeError):
-        fx_wand.border_color = 0xDEADBEEF
+def test_set_get_border_color():
+    with Drawing() as ctx:
+        with Color("#0F0") as green:
+            ctx.border_color = green
+            assert green == ctx.border_color
+        ctx.border_color = 'orange'
+        assert ctx.border_color == Color('orange')
+        # Assert user error
+        with raises(TypeError):
+            ctx.border_color = 0xDEADBEEF
 
 
-def test_set_get_clip_path(fx_wand):
-    fx_wand.clip_path = 'path_id'
-    assert fx_wand.clip_path == 'path_id'
-    # Assert user error
-    with raises(TypeError):
-        fx_wand.clip_path = 0xDEADBEEF
+def test_set_get_clip_path():
+    with Drawing() as ctx:
+        ctx.clip_path = 'path_id'
+        assert ctx.clip_path == 'path_id'
+        # Assert user error
+        with raises(TypeError):
+            ctx.clip_path = 0xDEADBEEF
 
 
-def test_set_get_clip_rule(fx_wand):
-    fx_wand.clip_rule = 'evenodd'
-    assert fx_wand.clip_rule == 'evenodd'
-    with raises(TypeError):
-        fx_wand.clip_rule = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.clip_rule = 'not-a-rule'
+def test_set_get_clip_rule():
+    with Drawing() as ctx:
+        ctx.clip_rule = 'evenodd'
+        assert ctx.clip_rule == 'evenodd'
+        with raises(TypeError):
+            ctx.clip_rule = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.clip_rule = 'not-a-rule'
 
 
-def test_set_get_clip_units(fx_wand):
-    fx_wand.clip_units = 'object_bounding_box'
-    assert fx_wand.clip_units == 'object_bounding_box'
-    with raises(TypeError):
-        fx_wand.clip_units = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.clip_units = 'not-a-clip_unit'
+def test_set_get_clip_units():
+    with Drawing() as ctx:
+        ctx.clip_units = 'object_bounding_box'
+        assert ctx.clip_units == 'object_bounding_box'
+        with raises(TypeError):
+            ctx.clip_units = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.clip_units = 'not-a-clip_unit'
 
 
-def test_set_get_font(fx_wand, fx_asset):
-    fx_wand.font = str(fx_asset.join('League_Gothic.otf'))
-    assert fx_wand.font == str(fx_asset.join('League_Gothic.otf'))
-    with raises(TypeError):
-        fx_wand.font = 0xDEADBEEF
+def test_set_get_font(fx_asset):
+    with Drawing() as ctx:
+        ctx.font = str(fx_asset.join('League_Gothic.otf'))
+        assert ctx.font == str(fx_asset.join('League_Gothic.otf'))
+        with raises(TypeError):
+            ctx.font = 0xDEADBEEF
 
 
-def test_set_get_font_family(fx_wand):
-    assert fx_wand.font_family is None
-    fx_wand.font_family = 'sans-serif'
-    assert fx_wand.font_family == 'sans-serif'
-    with raises(TypeError):
-        fx_wand.font_family = 0xDEADBEEF
+def test_set_get_font_family():
+    with Drawing() as ctx:
+        assert ctx.font_family is None
+        ctx.font_family = 'sans-serif'
+        assert ctx.font_family == 'sans-serif'
+        with raises(TypeError):
+            ctx.font_family = 0xDEADBEEF
 
 
-def test_set_get_font_resolution(fx_wand):
-    fx_wand.font_resolution = (78.0, 78.0)
-    assert fx_wand.font_resolution == (78.0, 78.0)
-    with raises(TypeError):
-        fx_wand.font_resolution = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.font_resolution = (78.0, 78.0, 78.0)
+def test_set_get_font_resolution():
+    with Drawing() as ctx:
+        ctx.font_resolution = (78.0, 78.0)
+        assert ctx.font_resolution == (78.0, 78.0)
+        with raises(TypeError):
+            ctx.font_resolution = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.font_resolution = (78.0, 78.0, 78.0)
 
 
-def test_set_get_font_size(fx_wand):
-    fx_wand.font_size = 22.2
-    assert fx_wand.font_size == 22.2
-    with raises(TypeError):
-        fx_wand.font_size = '22.2%'
-    with raises(ValueError):
-        fx_wand.font_size = -22.2
+def test_set_get_font_size():
+    with Drawing() as ctx:
+        ctx.font_size = 22.2
+        assert ctx.font_size == 22.2
+        with raises(TypeError):
+            ctx.font_size = '22.2%'
+        with raises(ValueError):
+            ctx.font_size = -22.2
 
 
-def test_set_get_font_stretch(fx_wand):
-    fx_wand.font_stretch = 'condensed'
-    assert fx_wand.font_stretch == 'condensed'
-    with raises(TypeError):
-        fx_wand.font_stretch = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.font_stretch = 'not-a-stretch-type'
+def test_set_get_font_stretch():
+    with Drawing() as ctx:
+        ctx.font_stretch = 'condensed'
+        assert ctx.font_stretch == 'condensed'
+        with raises(TypeError):
+            ctx.font_stretch = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.font_stretch = 'not-a-stretch-type'
 
 
-def test_set_get_font_style(fx_wand):
-    fx_wand.font_style = 'italic'
-    assert fx_wand.font_style == 'italic'
-    with raises(TypeError):
-        fx_wand.font_style = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.font_style = 'not-a-style-type'
+def test_set_get_font_style():
+    with Drawing() as ctx:
+        ctx.font_style = 'italic'
+        assert ctx.font_style == 'italic'
+        with raises(TypeError):
+            ctx.font_style = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.font_style = 'not-a-style-type'
 
 
-def test_set_get_font_weight(fx_wand):
-    fx_wand.font_weight = 400  # Normal
-    assert fx_wand.font_weight == 400
-    with raises(TypeError):
-        fx_wand.font_weight = '400'
+def test_set_get_font_weight():
+    with Drawing() as ctx:
+        ctx.font_weight = 400  # Normal
+        assert ctx.font_weight == 400
+        with raises(TypeError):
+            ctx.font_weight = '400'
 
 
-def test_set_get_fill_color(fx_wand):
-    with Color('#333333') as black:
-        fx_wand.fill_color = black
-    assert fx_wand.fill_color == Color('#333333')
-    fx_wand.fill_color = 'pink'
-    fx_wand.fill_color == Color('PINK')
+def test_set_get_fill_color():
+    with Drawing() as ctx:
+        with Color('#333333') as black:
+            ctx.fill_color = black
+        assert ctx.fill_color == Color('#333333')
+        ctx.fill_color = 'pink'
+        ctx.fill_color == Color('PINK')
 
 
-def test_set_get_stroke_color(fx_wand):
-    with Color('#333333') as black:
-        fx_wand.stroke_color = black
-    assert fx_wand.stroke_color == Color('#333333')
-    fx_wand.stroke_color = 'skyblue'
-    assert fx_wand.stroke_color == Color('SkyBlue')
+def test_set_get_stroke_color():
+    with Drawing() as ctx:
+        with Color('#333333') as black:
+            ctx.stroke_color = black
+        assert ctx.stroke_color == Color('#333333')
+        ctx.stroke_color = 'skyblue'
+        assert ctx.stroke_color == Color('SkyBlue')
 
 
-def test_set_get_stroke_width(fx_wand):
-    fx_wand.stroke_width = 5
-    assert fx_wand.stroke_width == 5
+def test_set_get_stroke_width():
+    with Drawing() as ctx:
+        ctx.stroke_width = 5
+        assert ctx.stroke_width == 5
 
 
-def test_set_get_text_alignment(fx_wand):
-    fx_wand.text_alignment = 'center'
-    assert fx_wand.text_alignment == 'center'
-    with raises(TypeError):
-        fx_wand.text_alignment = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.text_alignment = 'not-a-text-alignment-type'
+def test_set_get_text_alignment():
+    with Drawing() as ctx:
+        ctx.text_alignment = 'center'
+        assert ctx.text_alignment == 'center'
+        with raises(TypeError):
+            ctx.text_alignment = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.text_alignment = 'not-a-text-alignment-type'
 
 
-def test_set_get_text_antialias(fx_wand):
-    fx_wand.text_antialias = True
-    assert fx_wand.text_antialias is True
+def test_set_get_text_antialias():
+    with Drawing() as ctx:
+        ctx.text_antialias = True
+        assert ctx.text_antialias is True
 
 
-def test_set_get_text_decoration(fx_wand):
-    fx_wand.text_decoration = 'underline'
-    assert fx_wand.text_decoration == 'underline'
-    with raises(TypeError):
-        fx_wand.text_decoration = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.text_decoration = 'not-a-text-decoration-type'
+def test_set_get_text_decoration():
+    with Drawing() as ctx:
+        ctx.text_decoration = 'underline'
+        assert ctx.text_decoration == 'underline'
+        with raises(TypeError):
+            ctx.text_decoration = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.text_decoration = 'not-a-text-decoration-type'
 
 
-def test_set_get_text_direction(fx_wand):
-    try:
-        fx_wand.text_direction = 'right_to_left'
-        assert fx_wand.text_direction == 'right_to_left'
-    except WandLibraryVersionError:
-        skip("DrawGetTextDirection not supported by installed drawing library")
+@mark.skipif(MAGICK_VERSION_NUMBER < 0x689,
+             reason='DrawGetTextDirection not supported.')
+def test_set_get_text_direction():
+    with Drawing() as ctx:
+        ctx.text_direction = 'right_to_left'
+        assert ctx.text_direction == 'right_to_left'
 
 
-def test_set_get_text_encoding(fx_wand):
-    fx_wand.text_encoding = 'UTF-8'
-    assert fx_wand.text_encoding == 'UTF-8'
-    fx_wand.text_encoding = None
+def test_set_get_text_encoding():
+    with Drawing() as ctx:
+        ctx.text_encoding = 'UTF-8'
+        assert ctx.text_encoding == 'UTF-8'
+        ctx.text_encoding = None
 
 
-def test_set_get_text_interline_spacing(fx_wand):
-    fx_wand.text_interline_spacing = 10.11
-    assert fx_wand.text_interline_spacing == 10.11
-    with raises(TypeError):
-        fx_wand.text_interline_spacing = '10.11'
+def test_set_get_text_interline_spacing():
+    with Drawing() as ctx:
+        ctx.text_interline_spacing = 10.11
+        assert ctx.text_interline_spacing == 10.11
+        with raises(TypeError):
+            ctx.text_interline_spacing = '10.11'
 
 
-def test_set_get_text_interword_spacing(fx_wand):
-    fx_wand.text_interword_spacing = 5.55
-    assert fx_wand.text_interword_spacing == 5.55
-    with raises(TypeError):
-        fx_wand.text_interline_spacing = '5.55'
+def test_set_get_text_interword_spacing():
+    with Drawing() as ctx:
+        ctx.text_interword_spacing = 5.55
+        assert ctx.text_interword_spacing == 5.55
+        with raises(TypeError):
+            ctx.text_interline_spacing = '5.55'
 
 
-def test_set_get_text_kerning(fx_wand):
-    fx_wand.text_kerning = 10.22
-    assert fx_wand.text_kerning == 10.22
-    with raises(TypeError):
-        fx_wand.text_kerning = '10.22'
+def test_set_get_text_kerning():
+    with Drawing() as ctx:
+        ctx.text_kerning = 10.22
+        assert ctx.text_kerning == 10.22
+        with raises(TypeError):
+            ctx.text_kerning = '10.22'
 
 
-def test_set_get_text_under_color(fx_wand):
-    with Color('#333333') as black:
-        fx_wand.text_under_color = black
-    assert fx_wand.text_under_color == Color('#333333')
-    fx_wand.text_under_color = '#333'  # Smoke test
-    with raises(TypeError):
-        fx_wand.text_under_color = 0xDEADBEEF
+def test_set_get_text_under_color():
+    with Drawing() as ctx:
+        with Color('#333333') as black:
+            ctx.text_under_color = black
+        assert ctx.text_under_color == Color('#333333')
+        ctx.text_under_color = '#333'  # Smoke test
+        with raises(TypeError):
+            ctx.text_under_color = 0xDEADBEEF
 
 
-def test_set_get_vector_graphics(fx_wand):
-    fx_wand.stroke_width = 7
-    xml = fx_wand.vector_graphics
-    assert xml.index("<stroke-width>7</stroke-width>") > 0
-    fx_wand.vector_graphics = ('<wand><stroke-width>'
-                               '8</stroke-width></wand>')
-    xml = fx_wand.vector_graphics
-    assert xml.index("<stroke-width>8</stroke-width>") > 0
-    with raises(TypeError):
-        fx_wand.vector_graphics = 0xDEADBEEF
+def test_set_get_vector_graphics():
+    with Drawing() as ctx:
+        ctx.stroke_width = 7
+        xml = ctx.vector_graphics
+        assert xml.index("<stroke-width>7</stroke-width>") > 0
+        ctx.vector_graphics = '<wand><stroke-width>8</stroke-width></wand>'
+        xml = ctx.vector_graphics
+        assert xml.index("<stroke-width>8</stroke-width>") > 0
+        with raises(TypeError):
+            ctx.vector_graphics = 0xDEADBEEF
 
 
-def test_set_get_gravity(fx_wand):
-    fx_wand.gravity = 'center'
-    assert fx_wand.gravity == 'center'
-    with raises(TypeError):
-        fx_wand.gravity = 0xDEADBEEF
-    with raises(ValueError):
-        fx_wand.gravity = 'not-a-gravity-type'
+def test_set_get_gravity():
+    with Drawing() as ctx:
+        ctx.gravity = 'center'
+        assert ctx.gravity == 'center'
+        with raises(TypeError):
+            ctx.gravity = 0xDEADBEEF
+        with raises(ValueError):
+            ctx.gravity = 'not-a-gravity-type'
 
 
-def test_clone_drawing_wand(fx_wand):
-    fx_wand.text_kerning = 10.22
-    funcs = (lambda img: Drawing(drawing=fx_wand),
-             lambda img: fx_wand.clone())
-    for func in funcs:
-        with func(fx_wand) as cloned:
-            assert fx_wand.resource is not cloned.resource
-            assert fx_wand.text_kerning == cloned.text_kerning
+def test_clone_drawing_wand():
+    with Drawing() as ctx:
+        ctx.text_kerning = 10.22
+        funcs = (lambda img: Drawing(drawing=ctx),
+                 lambda img: ctx.clone())
+        for func in funcs:
+            with func(ctx) as cloned:
+                assert ctx.resource is not cloned.resource
+                assert ctx.text_kerning == cloned.text_kerning
 
 
-def test_clear_drawing_wand(fx_wand):
-    fx_wand.text_kerning = 10.22
-    assert fx_wand.text_kerning == 10.22
-    fx_wand.clear()
-    assert fx_wand.text_kerning == 0
+def test_clear_drawing_wand():
+    with Drawing() as ctx:
+        ctx.text_kerning = 10.22
+        assert ctx.text_kerning == 10.22
+        ctx.clear()
+        assert ctx.text_kerning == 0
 
 
 def test_composite():
@@ -286,14 +304,13 @@ def test_draw_circle(fx_asset):
 
 
 def test_draw_color():
-    white = Color('WHITE')
-    black = Color('BLACK')
-    with Image(width=50, height=50, background=white) as img:
-        with Drawing() as draw:
-            draw.fill_color = black
-            draw.color(25, 25, 'floodfill')
-            draw.draw(img)
-            assert img[25, 25] == black
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.fill_color = 'black'
+            ctx.color(25, 25, 'floodfill')
+            ctx.draw(img)
+        assert was != img.signature
 
 
 def test_draw_color_user_error():
@@ -306,35 +323,29 @@ def test_draw_color_user_error():
             draw.color(1, 2, 'apples')
 
 
-def test_draw_ellipse(fx_wand):
-    gray, red = Color('#ccc'), Color('#f00')
-    with Image(width=50, height=50, background=gray) as img:
-        with Drawing() as draw:
-            draw.fill_color = red
-            draw.ellipse((25, 25),  # origin
-                         (20, 10))  # radius
-            draw.draw(img)
-            assert img[25, 10] == gray
-            assert img[45, 25] == red
+def test_draw_ellipse():
+    with Image(width=50, height=50, background='#ccc') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.fill_color = 'red'
+            ctx.ellipse((25, 25),  # origin
+                        (20, 10))  # radius
+            ctx.draw(img)
+        assert was != img.signature
 
 
-def test_draw_line(fx_wand):
-    gray = Color('#ccc')
-    with Image(width=10, height=10, background=gray) as img:
-        with Color('#333333') as black:
-            fx_wand.fill_color = black
-        fx_wand.line((5, 5), (7, 5))
-        fx_wand.draw(img)
-        assert img[4, 5] == Color('#ccc')
-        assert img[5, 5] == Color('#333333')
-        assert img[6, 5] == Color('#333333')
-        assert img[7, 5] == Color('#333333')
-        assert img[8, 5] == Color('#ccc')
+def test_draw_line():
+    with Image(width=10, height=10, background='#ccc') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.fill_color = 'black'
+            ctx.line((5, 5), (7, 5))
+            ctx.draw(img)
+        assert was != img.signature
 
 
-@mark.xfail(MAGICK_VERSION_NUMBER >= 0x700,
-            reason='wand.drawing.Drawing.matte removed with IM 7.',
-            raises=AttributeError)
+@mark.skipif(MAGICK_VERSION_NUMBER >= 0x700,
+             reason='wand.drawing.Drawing.matte removed with IM 7.')
 def test_draw_matte():
     white = Color('rgba(0, 255, 255, 5%)')
     transparent = Color('transparent')
@@ -346,9 +357,8 @@ def test_draw_matte():
             assert img[25, 25] == transparent
 
 
-@mark.xfail(MAGICK_VERSION_NUMBER >= 0x700,
-            reason='wand.drawing.Drawing.matte removed with IM 7.',
-            raises=AttributeError)
+@mark.skipif(MAGICK_VERSION_NUMBER >= 0x700,
+             reason='wand.drawing.Drawing.matte removed with IM 7.')
 def test_draw_matte_user_error():
     with Drawing() as draw:
         with raises(TypeError):
@@ -359,9 +369,8 @@ def test_draw_matte_user_error():
             draw.matte(1, 2, 'apples')
 
 
-@mark.xfail(MAGICK_VERSION_NUMBER < 0x700,
-            reason='wand.drawing.Drawing.alpha was added with IM 7.',
-            raises=AttributeError)
+@mark.skipif(MAGICK_VERSION_NUMBER < 0x700,
+             reason='wand.drawing.Drawing.alpha was added with IM 7.')
 def test_draw_alpha():
     transparent = Color('transparent')
     with Image(width=50, height=50, pseudo='xc:white') as img:
@@ -372,9 +381,8 @@ def test_draw_alpha():
         assert img[25, 25] == transparent
 
 
-@mark.xfail(MAGICK_VERSION_NUMBER < 0x700,
-            reason='wand.drawing.Drawing.alpha was added with IM 7.',
-            raises=AttributeError)
+@mark.skipif(MAGICK_VERSION_NUMBER < 0x700,
+             reason='wand.drawing.Drawing.alpha was added with IM 7.')
 def test_draw_alpha_user_error():
     with Drawing() as draw:
         with raises(TypeError):
@@ -396,7 +404,7 @@ def test_draw_point():
             assert img[2, 2] == black
 
 
-def test_draw_polygon(fx_wand):
+def test_draw_polygon():
     white = Color('WHITE')
     red = Color('RED')
     blue = Color('BLUE')
@@ -413,7 +421,7 @@ def test_draw_polygon(fx_wand):
             assert img[35, 15] == img[35, 35] == white
 
 
-def test_draw_polyline(fx_wand):
+def test_draw_polyline():
     white = Color('WHITE')
     red = Color('RED')
     blue = Color('BLUE')
@@ -437,7 +445,7 @@ def test_draw_push_pop():
         assert 2 == draw.stroke_width
 
 
-def test_draw_bezier(fx_wand):
+def test_draw_bezier():
     white = Color('WHITE')
     red = Color('RED')
     blue = Color('BLUE')
@@ -613,20 +621,19 @@ def test_draw_move_user_error():
     [('right', 40), ('width', 30)],
     [('bottom', 40), ('height', 30)]
 ))
-def test_draw_rectangle(kwargs, fx_wand):
+def test_draw_rectangle(kwargs):
     white = Color('WHITE')
     black = Color('BLACK')
     gray = Color('#ccc')
     with Image(width=50, height=50, background=white) as img:
-        fx_wand.stroke_width = 2
-        fx_wand.fill_color = black
-        fx_wand.stroke_color = gray
-        fx_wand.rectangle(left=10, top=10, **dict(kwargs))
-        fx_wand.draw(img)
-        assert (img[7, 7] == img[7, 42] == img[42, 7] ==
-                img[42, 42] == img[0, 0] == img[49, 49] == white)
-        assert (img[12, 12] == img[12, 38] == img[38, 12] ==
-                img[38, 38] == black)
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.stroke_width = 2
+            ctx.fill_color = black
+            ctx.stroke_color = gray
+            ctx.rectangle(left=10, top=10, **dict(kwargs))
+            ctx.draw(img)
+        assert was != img.signature
 
 
 @mark.parametrize('kwargs', itertools.product(
@@ -635,115 +642,104 @@ def test_draw_rectangle(kwargs, fx_wand):
     [('yradius', 20)],
     [('radius', 10)]
 ))
-def test_draw_rectangle_with_radius(kwargs, fx_wand):
-    white = Color('WHITE')
-    black = Color('BLACK')
-    gray = Color('#ccc')
-    with Image(width=50, height=50, background=white) as img:
-        fx_wand.stroke_width = 2
-        fx_wand.fill_color = black
-        fx_wand.stroke_color = gray
-        fx_wand.rectangle(left=10, top=10,
+def test_draw_rectangle_with_radius(kwargs):
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.stroke_width = 2
+            ctx.fill_color = 'black'
+            ctx.stroke_color = '#ccc'
+            ctx.rectangle(left=10, top=10,
                           width=30, height=30, **dict(kwargs))
-        fx_wand.draw(img)
-        assert img[10, 10] == img[40, 40] == white
-        assert img[26, 12] == img[26, 36] == black
+            ctx.draw(img)
+        assert was != img.signature
 
 
 def test_draw_rotate():
-    white = Color('WHITE')
-    black = Color('BLACK')
-    with Image(width=50, height=50, background=white) as img:
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
         with Drawing() as draw:
-            draw.stroke_color = black
+            draw.stroke_color = 'black'
             draw.rotate(45)
             draw.line((3, 3), (35, 35))
             draw.draw(img)
-            assert img[0, 49] == black
+        assert was != img.signature
 
 
-def test_draw_scale(fx_wand):
-    white = Color('WHITE')
-    black = Color('BLACK')
-    with Image(width=50, height=50, background=white) as img:
-        fx_wand.fill_color = black
-        fx_wand.scale(x=2.0, y=0.5)
-        fx_wand.rectangle(top=5, left=5, width=20, height=20)
-        fx_wand.draw(img)
-        # if width was scaled up by 200%
-        assert img[45, 10] == black
-        # if height was scaled down by 50%
-        assert img[20, 20] == white
+def test_draw_scale():
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.fill_color = 'black'
+            ctx.scale(x=2.0, y=0.5)
+            ctx.rectangle(top=5, left=5, width=20, height=20)
+            ctx.draw(img)
+        assert was != img.signature
 
 
-def test_set_fill_pattern_url(fx_wand):
-    white = Color('#fff')
-    green = Color('#0f0')
-    black = Color('#000')
-    with Image(width=50, height=50, background=white) as img:
-        fx_wand.push_pattern('green_circle', 0, 0, 10, 10)
-        fx_wand.fill_color = green
-        fx_wand.stroke_color = black
-        fx_wand.circle(origin=(5, 5), perimeter=(5, 0))
-        fx_wand.pop_pattern()
-        fx_wand.set_fill_pattern_url('#green_circle')
-        fx_wand.rectangle(top=5, left=5, width=40, height=40)
-        fx_wand.draw(img)
-        assert img[17, 17] == green
+def test_set_fill_pattern_url():
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.push_pattern('green_circle', 0, 0, 10, 10)
+            ctx.fill_color = 'green'
+            ctx.stroke_color = 'black'
+            ctx.circle(origin=(5, 5), perimeter=(5, 0))
+            ctx.pop_pattern()
+            ctx.set_fill_pattern_url('#green_circle')
+            ctx.rectangle(top=5, left=5, width=40, height=40)
+            ctx.draw(img)
+        assert was != img.signature
 
 
-def test_set_stroke_pattern_url(fx_wand):
-    white = Color('#fff')
-    green = Color('#0f0')
-    with Image(width=50, height=50, background=white) as img:
-        fx_wand.push_pattern('green_ring', 0, 0, 6, 6)
-        fx_wand.fill_color = green
-        fx_wand.stroke_color = white
-        fx_wand.circle(origin=(3, 3), perimeter=(3, 0))
-        fx_wand.pop_pattern()
-        fx_wand.set_stroke_pattern_url('#green_ring')
-        fx_wand.stroke_width = 6
-        fx_wand.rectangle(top=5, left=5, width=40, height=40)
-        fx_wand.draw(img)
-        assert img[45, 45] == green
+def test_set_stroke_pattern_url():
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.push_pattern('green_ring', 0, 0, 6, 6)
+            ctx.fill_color = 'green'
+            ctx.stroke_color = 'white'
+            ctx.circle(origin=(3, 3), perimeter=(3, 0))
+            ctx.pop_pattern()
+            ctx.set_stroke_pattern_url('#green_ring')
+            ctx.stroke_width = 6
+            ctx.rectangle(top=5, left=5, width=40, height=40)
+            ctx.draw(img)
+        assert was != img.signature
 
 
 def test_draw_skew():
-    white = Color('#fff')
-    black = Color('#000')
-    with Image(width=50, height=50, background=white) as img:
-        with Drawing() as draw:
-            draw.stroke_color = black
-            draw.skew(x=11, y=-24)
-            draw.line((3, 3), (35, 35))
-            draw.draw(img)
-            assert img[43, 42] == black
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.stroke_color = 'black'
+            ctx.skew(x=11, y=-24)
+            ctx.line((3, 3), (35, 35))
+            ctx.draw(img)
+        assert was != img.signature
 
 
 def test_draw_translate():
-    white = Color('#fff')
-    black = Color('#000')
-    with Image(width=50, height=50, background=white) as img:
-        with Drawing() as draw:
-            draw.stroke_color = black
-            draw.translate(x=5, y=5)
-            draw.line((3, 3), (35, 35))
-            draw.draw(img)
-            assert img[40, 40] == black
+    with Image(width=50, height=50, background='white') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.stroke_color = 'black'
+            ctx.translate(x=5, y=5)
+            ctx.line((3, 3), (35, 35))
+            ctx.draw(img)
+        assert was != img.signature
 
 
 def test_draw_text(fx_asset):
-    white = Color('#fff')
-    with Image(width=100, height=100, background=white) as img:
+    with Image(width=100, height=100, background='white') as img:
         was = img.signature
-        with Drawing() as draw:
-            draw.font = str(fx_asset.join('League_Gothic.otf'))
-            draw.font_size = 25
-            with Color('#000') as bk:
-                draw.fill_color = bk
-            draw.gravity = 'west'
-            draw.text(0, 0, 'Hello Wand')
-            draw.draw(img)
+        with Drawing() as ctx:
+            ctx.font = str(fx_asset.join('League_Gothic.otf'))
+            ctx.font_size = 25
+            ctx.fill_color = 'black'
+            ctx.gravity = 'west'
+            ctx.text(0, 0, 'Hello Wand')
+            ctx.draw(img)
         assert was != img.signature
 
 
@@ -793,142 +789,159 @@ def test_regression_issue_163(tmpdir):
             image.save(filename=str(tmpdir.join('out.jpg')))
 
 
-def test_set_get_fill_opacity(fx_wand):
-    fx_wand.fill_opacity = 1.0
-    assert fx_wand.fill_opacity == 1.0
+def test_set_get_fill_opacity():
+    with Drawing() as ctx:
+        ctx.fill_opacity = 1.0
+        assert ctx.fill_opacity == 1.0
 
 
-def test_set_get_fill_opacity_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.fill_opacity = "1.5"
+def test_set_get_fill_opacity_user_error():
+    with Drawing() as ctx:
+        with raises(TypeError):
+            ctx.fill_opacity = "1.5"
 
 
-def test_set_get_fill_rule(fx_wand):
-    valid = 'evenodd'
-    notvalid = 'error'
-    invalid = (1, 2)
-    fx_wand.fill_rule = valid
-    assert fx_wand.fill_rule == valid
-    with raises(ValueError):
-        fx_wand.fill_rule = notvalid
-    with raises(TypeError):
-        fx_wand.fill_rule = invalid
-    fx_wand.fill_rule = 'undefined'  # reset
+def test_set_get_fill_rule():
+    with Drawing() as ctx:
+        valid = 'evenodd'
+        notvalid = 'error'
+        invalid = (1, 2)
+        ctx.fill_rule = valid
+        assert ctx.fill_rule == valid
+        with raises(ValueError):
+            ctx.fill_rule = notvalid
+        with raises(TypeError):
+            ctx.fill_rule = invalid
+        ctx.fill_rule = 'undefined'  # reset
 
 
 @mark.skipif(MAGICK_VERSION_NUMBER < 0x700,
              reason='DrawGetOpacity always returns 1.0')
-def test_set_get_opacity(fx_wand):
-    assert fx_wand.opacity == 1.0
-    fx_wand.push()
-    fx_wand.opacity = 0.5
-    fx_wand.push()
-    fx_wand.opacity = 0.25
-    assert 0.24 < fx_wand.opacity < 0.26  # Expect float precision issues
-    fx_wand.pop()
-    assert 0.49 < fx_wand.opacity < 0.51  # Expect float precision issues
-    fx_wand.pop()
+def test_set_get_opacity():
+    with Drawing() as ctx:
+        assert ctx.opacity == 1.0
+        ctx.push()
+        ctx.opacity = 0.5
+        ctx.push()
+        ctx.opacity = 0.25
+        assert 0.24 < ctx.opacity < 0.26  # Expect float precision issues
+        ctx.pop()
+        assert 0.49 < ctx.opacity < 0.51  # Expect float precision issues
+        ctx.pop()
 
 
-def test_set_get_stroke_antialias(fx_wand):
-    fx_wand.stroke_antialias = False
-    assert not fx_wand.stroke_antialias
+def test_set_get_stroke_antialias():
+    with Drawing() as ctx:
+        ctx.stroke_antialias = False
+        assert not ctx.stroke_antialias
 
 
-def test_set_get_stroke_dash_array(fx_wand):
-    dash_array = [2, 1, 4, 1]
-    fx_wand.stroke_dash_array = dash_array
-    assert fx_wand.stroke_dash_array == dash_array
+def test_set_get_stroke_dash_array():
+    with Drawing() as ctx:
+        dash_array = [2, 1, 4, 1]
+        ctx.stroke_dash_array = dash_array
+        assert ctx.stroke_dash_array == dash_array
 
 
-def test_set_get_stroke_dash_offset(fx_wand):
-    fx_wand.stroke_dash_offset = 0.5
-    assert fx_wand.stroke_dash_offset == 0.5
+def test_set_get_stroke_dash_offset():
+    with Drawing() as ctx:
+        ctx.stroke_dash_offset = 0.5
+        assert ctx.stroke_dash_offset == 0.5
 
 
-def test_set_get_stroke_line_cap(fx_wand):
-    fx_wand.stroke_line_cap = 'round'
-    assert fx_wand.stroke_line_cap == 'round'
+def test_set_get_stroke_line_cap():
+    with Drawing() as ctx:
+        ctx.stroke_line_cap = 'round'
+        assert ctx.stroke_line_cap == 'round'
 
 
-def test_set_get_stroke_line_cap_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.stroke_line_cap = 0x74321870
-    with raises(ValueError):
-        fx_wand.stroke_line_cap = 'apples'
-
-
-def test_set_get_stroke_line_join(fx_wand):
-    fx_wand.stroke_line_join = 'miter'
-    assert fx_wand.stroke_line_join == 'miter'
-
-
-def test_set_get_stroke_line_join_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.stroke_line_join = 0x74321870
-    with raises(ValueError):
-        fx_wand.stroke_line_join = 'apples'
-
-
-def test_set_get_stroke_miter_limit(fx_wand):
-    fx_wand.stroke_miter_limit = 5
-    assert fx_wand.stroke_miter_limit == 5
-
-
-def test_set_get_stroke_miter_limit_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.stroke_miter_limit = '5'
-
-
-def test_set_get_stroke_opacity(fx_wand):
-    fx_wand.stroke_opacity = 1.0
-    assert fx_wand.stroke_opacity == 1.0
-
-
-def test_set_get_stroke_opacity_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.stroke_opacity = '1.0'
-
-
-def test_set_get_stroke_width_user_error(fx_wand):
-    with raises(TypeError):
-        fx_wand.stroke_width = '0.1234'
-    with raises(ValueError):
-        fx_wand.stroke_width = -1.5
-
-
-def test_draw_affine(fx_wand):
-    skyblue = Color('skyblue')
-    black = Color('black')
-    with Image(width=100, height=100, background=skyblue) as img:
-        img.format = 'png'
-        fx_wand.affine([1.5, 0.5, 0, 1.5, 45, 25])
-        fx_wand.rectangle(top=5, left=5, width=25, height=25)
-        fx_wand.draw(img)
-        assert img[25, 25] == skyblue
-        assert img[75, 75] == black
-        with raises(ValueError):
-            fx_wand.affine([1.0])
+def test_set_get_stroke_line_cap_user_error():
+    with Drawing() as ctx:
         with raises(TypeError):
-            fx_wand.affine(['a', 'b', 'c', 'd', 'e', 'f'])
+            ctx.stroke_line_cap = 0x74321870
+        with raises(ValueError):
+            ctx.stroke_line_cap = 'apples'
 
 
-def test_draw_clip_path(fx_wand):
+def test_set_get_stroke_line_join():
+    with Drawing() as ctx:
+        ctx.stroke_line_join = 'miter'
+        assert ctx.stroke_line_join == 'miter'
+
+
+def test_set_get_stroke_line_join_user_error():
+    with Drawing() as ctx:
+        with raises(TypeError):
+            ctx.stroke_line_join = 0x74321870
+        with raises(ValueError):
+            ctx.stroke_line_join = 'apples'
+
+
+def test_set_get_stroke_miter_limit():
+    with Drawing() as ctx:
+        ctx.stroke_miter_limit = 5
+        assert ctx.stroke_miter_limit == 5
+
+
+def test_set_get_stroke_miter_limit_user_error():
+    with Drawing() as ctx:
+        with raises(TypeError):
+            ctx.stroke_miter_limit = '5'
+
+
+def test_set_get_stroke_opacity():
+    with Drawing() as ctx:
+        ctx.stroke_opacity = 1.0
+        assert ctx.stroke_opacity == 1.0
+
+
+def test_set_get_stroke_opacity_user_error():
+    with Drawing() as ctx:
+        with raises(TypeError):
+            ctx.stroke_opacity = '1.0'
+
+
+def test_set_get_stroke_width_user_error():
+    with Drawing() as ctx:
+        with raises(TypeError):
+            ctx.stroke_width = '0.1234'
+        with raises(ValueError):
+            ctx.stroke_width = -1.5
+
+
+def test_draw_affine():
+    with Image(width=100, height=100, background='skyblue') as img:
+        was = img.signature
+        img.format = 'png'
+        with Drawing() as ctx:
+            ctx.affine([1.5, 0.5, 0, 1.5, 45, 25])
+            ctx.rectangle(top=5, left=5, width=25, height=25)
+            ctx.draw(img)
+        assert was != img.signature
+    with raises(ValueError):
+        with Drawing() as ctx:
+            ctx.affine([1.0])
+    with raises(TypeError):
+        with Drawing() as ctx:
+            ctx.affine(['a', 'b', 'c', 'd', 'e', 'f'])
+
+
+def test_draw_clip_path():
     skyblue = Color('skyblue')
     orange = Color('orange')
-    with Image(width=100, height=100, background=skyblue) as img:
-        fx_wand.push_defs()
-        fx_wand.push_clip_path("eyes_only")
-        fx_wand.push()
-        fx_wand.rectangle(top=0, left=0, width=50, height=50)
-        fx_wand.pop()
-        fx_wand.pop_clip_path()
-        fx_wand.pop_defs()
-
-        fx_wand.clip_path = "eyes_only"
-        fx_wand.clip_rule = "nonzero"
-        fx_wand.clip_path_units = "object_bounding_box"
-        fx_wand.fill_color = orange
-        fx_wand.rectangle(top=5, left=5, width=90, height=90)
-        fx_wand.draw(img)
+    with Image(width=100, height=100, background='skyblue') as img:
+        with Drawing() as ctx:
+            ctx.push_defs()
+            ctx.push_clip_path("eyes_only")
+            ctx.push()
+            ctx.rectangle(top=0, left=0, width=50, height=50)
+            ctx.pop()
+            ctx.pop_clip_path()
+            ctx.pop_defs()
+            ctx.clip_path = "eyes_only"
+            ctx.clip_rule = "nonzero"
+            ctx.clip_path_units = "object_bounding_box"
+            ctx.fill_color = orange
+            ctx.rectangle(top=5, left=5, width=90, height=90)
+            ctx.draw(img)
         assert img[75, 75] == skyblue